Move public-facing pages + data to frontend/
