+++ .github/workflows/deploy.yml
@@ -30,6 +30,9 @@ jobs:
         with:
           node-version: ${{ matrix.node-version }}
 
+
+      - run: cd frontend
+
       - run: npm install
 
       - run: npm run build
@@ -41,5 +44,5 @@ jobs:
         with:
           token: ${{ secrets.GITHUB_TOKEN }}
           branch: gh-pages
-          folder: out/
+          folder: frontend/out/
           clean: true
