+++ frontend/data/navigation.ts
