+++ frontend/data/projects.ts
