+++ frontend/data/publications.ts
