+++ frontend/interfaces/navigation.ts
