+++ frontend/interfaces/projects.ts
