+++ frontend/interfaces/publications.ts
